don't delete shared/boot parts in deleteAll (#1183880)
This is one potential approach to help with #1183880 . It
changes the deleteAll behaviour as follows:

* Partitions known to be shared with other OSes won't be deleted
* Boot partitions won't be deleted if there are unknown parts

You can delete these partitions only by specifically targeting
them. It also adds a variant of the ConfirmDeleteDialog for
boot partitions, with text specifically explaining that they
might be needed for other OSes to boot.

+++ pyanaconda/ui/gui/spokes/custom.py
@@ -69,6 +69,7 @@ from blivet.errors import NoDisksError
 from blivet.errors import NotEnoughFreeSpaceError
 from blivet.devicelibs import raid, crypto
 from blivet.devices import LUKSDevice
+from blivet.platform import platform
 
 from pyanaconda.storage_utils import ui_storage_logger, device_type_from_autopart
 from pyanaconda.storage_utils import DEVICE_TEXT_PARTITION, DEVICE_TEXT_MAP, DEVICE_TEXT_MD
@@ -1952,9 +1953,12 @@ class CustomPartitioningSpoke(NormalSpoke, StorageChecker):
             # on it.  Thus, we first need to confirm with the user and then
             # schedule actions to delete the thing.
             dialog = ConfirmDeleteDialog(self.data)
+            protected = platform.bootStage1ConstraintDict["format_types"]
+            bootpart = device.format.type in protected
             snapshots = (device.direct and not device.isleaf)
             dialog.refresh(getattr(device.format, "mountpoint", ""),
-                           device.name, root_name, snapshots=snapshots)
+                           device.name, root_name, snapshots=snapshots,
+                           bootpart=bootpart)
             with self.main_window.enlightbox(dialog.window):
                 rc = dialog.run()
 
@@ -1963,16 +1967,33 @@ class CustomPartitioningSpoke(NormalSpoke, StorageChecker):
                     return
 
             if dialog.deleteAll:
-                for dev in (s._device for s in page.members):
-                    self._destroy_device(dev)
-            else:
-                self._destroy_device(device)
+                otherpgs = (pg for pg in self._accordion.allPages if
+                            pg.pageTitle != page.pageTitle)
+                otherdevs = []
+                for otherpg in otherpgs:
+                    otherdevs.extend([mem._device.id for mem in otherpg.members])
+                # we never want to delete known-shared devs here. we also
+                # exclude 'device' as we'll unconditionally delete it later
+                for dev in (s._device for s in page.members if
+                            s._device.id not in otherdevs and
+                            s._device.id != device.id):
+                    # we only want to delete boot partitions if they're not
+                    # shared *and* we have no unknown partitions
+                    if (not self.unusedDevices or dev.format.type not in protected):
+                        log.debug("deleteall: removed %s", dev.name)
+                        self._destroy_device(dev)
+                    else:
+                        log.debug("deleteall: didn't remove %s", dev.name)
+
+            # We do this even in deleteAll case, see above comment
+            self._destroy_device(device)
 
         log.info("ui: removed device %s", device.name)
 
         # Now that devices have been removed from the installation root,
         # refreshing the display will have the effect of making them disappear.
         # It's like they never existed.
+        self._storage_playground.roots = findExistingInstallations(self._storage_playground.devicetree)
         self._updateSpaceDisplay()
         self._do_refresh()
